+++ packages/reactivity-core/index.ts
@@ -45,6 +45,6 @@ export {
     isReactive
 } from "./ReactiveImpl";
 export { syncEffect, syncEffectOnce, syncWatch, syncWatchValue } from "./sync";
-export { effect, watch, watchValue } from "./async";
+export { effect, watch, watchValue, nextTick } from "./async";
 export * from "./collections";
 export * from "./struct";