+++ packages/reactivity-core/CHANGELOG.md
@@ -1,6 +1,6 @@
 # @conterra/reactivity-core
 
-## v0.5.0 (unreleased)
+## v0.4.1 (unreleased)
 
 Improved `watch` ergonomics: